+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 
 * Dropped support for Django<1.8, django CMS<3.2.
 * Added liveblog application.
+* Refactored plugin filters: by default only current site are now shown.
 * Added standalone documentation at https://djangocms-blog.readthedocs.io.
 
 ******************